requirements.txt
streamlit==1.32.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.14
langchain-experimental==0.0.43
langchain-core==0.1.18
openai==1.12.0
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
pydantic==2.5.3
beautifulsoup4==4.12.2
requests==2.31.0
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
chromadb==0.4.22
python-decouple==3.8
tiktoken==0.5.2
duckduckgo-search==7.5.5
langsmith>=0.0.77,<0.1.0
